GCService: fixing a bug in the GCAllocate

+++ runtime/gc/heap.h
@@ -748,7 +748,7 @@ class Heap {
                                 uint64_t gc_duration,
                                 double adjusted_resize_factor,
                                 size_t* adjusted_max_free,
-                                size_t conc_latency);
+                                double conc_latency);
 
   size_t GetPercentFree();
